Add files via upload
In the last section, an array of values as input is provided to the model (Amount of Nitrogen, Phosphorus .... present in the environment), and the model predicts/recommends the appropriate crop to cultivate for optimized yield. So, new input should be given in form of array as shown with the sample in the last secction of this model.